requirements.txt
annotated-types==0.7.0
anyio==4.11.0
audioread==3.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
elevenlabs==2.21.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
Gammatone==1.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.8.4
mpmath==1.3.0
msgpack==1.1.2
networkx==3.2.1
numba==0.62.1
numpy==1.26.4
packaging==25.0
pandas==2.2.3
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.12.2
six==1.17.0
sniffio==1.3.1
snn_reservoir_py==2.0.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
